chore(nodejs): extract transport from sender, add back support for core http
