+++ test/util/proxy.ts
@@ -1,17 +1,19 @@
-import { Socket } from "node:net";
+import net, { Socket } from "node:net";
+import tls from "node:tls";
 import { write, listen, shutdown, connect, close } from "./proxyfunctions";
 
 // handles only a single client
 // client -> server (Proxy) -> remote (QuestDB)
 // client <- server (Proxy) <- remote (QuestDB)
 class Proxy {
-  client: unknown;
+  client: Socket;
   remote: Socket;
+  server: net.Server | tls.Server
 
   constructor() {
     this.remote = new Socket();
 
-    this.remote.on("data", async (data: unknown) => {
+    this.remote.on("data", async (data: string) => {
       console.info(`received from remote, forwarding to client: ${data}`);
       await write(this.client, data);
     });
@@ -25,14 +27,14 @@ class Proxy {
     });
   }
 
-  async start(listenPort: unknown, remotePort: unknown, remoteHost: unknown, tlsOptions: Record<string, unknown>) {
+  async start(listenPort: number, remotePort: number, remoteHost: string, tlsOptions: Record<string, unknown>) {
     return new Promise<void>((resolve) => {
       this.remote.on("ready", async () => {
         console.info("remote connection ready");
         await listen(
           this,
           listenPort,
-          async (data: unknown) => {
+          async (data: string) => {
             console.info(`received from client, forwarding to remote: ${data}`);
             await write(this.remote, data);
           },
